+++ file_test.go
@@ -1639,3 +1639,103 @@ func TestFlattenFile_FileHeaderError(t *testing.T) {
 		}
 	}
 }
+
+// TestFile_FlattenFileOneIATBatchHeader
+func TestFile_FlattenFileOneIATBatchHeader(t *testing.T) {
+	// open a file for reading. Any io.Reader Can be used
+	f, err := os.Open(filepath.Join("test", "testdata", "flattenIATBatchesOneBatchHeader.ach"))
+
+	if err != nil {
+		t.Fatal(err)
+	}
+	r := NewReader(f)
+	achFile, err := r.Read()
+	if err != nil {
+		t.Fatalf("Issue reading file: %+v \n", err)
+	}
+
+	of, err := achFile.FlattenBatches()
+
+	if err != nil {
+		t.Fatalf("Could not flatten the file: %+v \n", err)
+	}
+
+	if err := of.Validate(); err != nil {
+		t.Fatalf("Flattend file did not validate: %+v \n", err)
+	}
+}
+
+// TestFileFlattenFileMultipleIATBatchHeaders
+func TestFileFlattenFileMultipleIATBatchHeaders(t *testing.T) {
+	// open a file for reading. Any io.Reader Can be used
+	f, err := os.Open(filepath.Join("test", "testdata", "flattenIATBatchesMultipleBatchHeaders.ach"))
+
+	if err != nil {
+		t.Fatal(err)
+	}
+	r := NewReader(f)
+	achFile, err := r.Read()
+	if err != nil {
+		t.Fatalf("Issue reading file: %+v \n", err)
+	}
+
+	of, err := achFile.FlattenBatches()
+
+	if err != nil {
+		t.Fatalf("Could not flatten the file: %+v \n", err)
+	}
+
+	if err := of.Validate(); err != nil {
+		t.Fatalf("Flattend file did not validate: %+v \n", err)
+	}
+}
+
+// TestFile_FlattenFileOneADVBatchHeader
+func TestFile_FlattenFileOneADVBatchHeader(t *testing.T) {
+	// open a file for reading. Any io.Reader Can be used
+	f, err := os.Open(filepath.Join("test", "testdata", "flattenADVBatchesOneBatchHeader.ach"))
+
+	if err != nil {
+		t.Fatal(err)
+	}
+	r := NewReader(f)
+	achFile, err := r.Read()
+	if err != nil {
+		t.Fatalf("Issue reading file: %+v \n", err)
+	}
+
+	of, err := achFile.FlattenBatches()
+
+	if err != nil {
+		t.Fatalf("Could not flatten the file: %+v \n", err)
+	}
+
+	if err := of.Validate(); err != nil {
+		t.Fatalf("Flattend file did not validate: %+v \n", err)
+	}
+}
+
+// TestFileFlattenFileMultipleADVBatchHeaders
+func TestFileFlattenFileMultipleADVTBatchHeaders(t *testing.T) {
+	// open a file for reading. Any io.Reader Can be used
+	f, err := os.Open(filepath.Join("test", "testdata", "flattenADVBatchesMultipleBatchHeaders.ach"))
+
+	if err != nil {
+		t.Fatal(err)
+	}
+	r := NewReader(f)
+	achFile, err := r.Read()
+	if err != nil {
+		t.Fatalf("Issue reading file: %+v \n", err)
+	}
+
+	of, err := achFile.FlattenBatches()
+
+	if err != nil {
+		t.Fatalf("Could not flatten the file: %+v \n", err)
+	}
+
+	if err := of.Validate(); err != nil {
+		t.Fatalf("Flattend file did not validate: %+v \n", err)
+	}
+}